Fix dlsym_default to not rely on dlsym and dlopen
The old DLSYM_DEFAULT implementation works exactly as expected
when given an actual handle for a library. However, potential
problems arose with the use of the RTLD_DEFAULT and RTLD_NEXT
pseudo-handles. In these cases, the function relied on calls
to dlopen and dlsym which could cause an infinite loop if called
before the complete initialization of libc.
This commit replaces the usage of dlopen and dlsym with dladdr1 which
simplifies the process greatly and should fix this problem.